requirements.txt
asgiref==3.9.1
async-timeout==5.0.1
Django==5.2.4
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-phonenumber-field==8.1.0
django-redis==6.0.0
django-rest-knox==5.0.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
inflection==0.5.1
packaging==25.0
phonenumbers==9.0.10
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
sqlparse==0.5.3
typing_extensions==4.14.1
uritemplate==4.2.0
whitenoise==6.9.0
